****The old PR #19574 works without problems, so I close this one.**** 

This PR does not cause accuracy drop.

| Models accuracy| Before Acc | After Acc |
| ------------- |:-------------:| -----:|
| Resnet50      |  0.7654 (full bin)  |   0.7654 (full bin)|
|mobilenetv1   |       0.7000       |     0.7000             |
|mobilenetv2  |        0.7600      |         0.7600         |
|googlenet|      0.7100        |            0.7100      |
|resnet101 |     0.8100         |       0.8100           |
|vgg16|     0.7500           |             0.7500     |
|vgg19|       0.7400       |            0.7400      |
|mobilenet_ssd |       0.7508       |      0.7508           |
	
| Performance| Before [FPS]| After [FPS]|
| ------------- |:-------------:| -----:|
| Resnet50      |  250.5979  (full bin)   |    252.6608 (full bin)   |
|mobilenetv1   |     483.5460 ~ 609.6494         |    494.3980              | 
|mobilenetv2  |      513.0954        |    577.6745              |
|googlenet|      325.2685        |       325.6318           |
|resnet101 |        128.3156      |       122.7690           |
|vgg16|     61.9907         |            69.7824      |
|vgg19|      54.3106        |         55.8488         |
|mobilenet_ssd |     286.2319         |      292.7095           |