@lidanqing-intel 2019-09-06
This is a reopen PR for https://github.com/PaddlePaddle/Paddle/pull/19574

Because  https://github.com/PaddlePaddle/Paddle/pull/19574 was from this Monday. During these days, one PR was merged and caused conflict. I rebased but I assume merging this PR will still require solving conflicts. Hence I make a new PR, the code is the same as the old. Only difference is I add `mask_bias_reorder` in second commit.

Hi@luotao, after this PR, INT8 models has same accuracy as before. I also changed things according to @Sand3r- reviews. Could you please review too.